ai-pothole-child/requirements.txt
torch>=2.0.0
fastapi
uvicorn
pydantic
numpy
ultralytics
transformers
pillow
requests
python-multipart
